requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-bootstrap3==21.2
django-polymorphic==3.1.0
idna==3.3
Naked==0.1.31
psycopg2==2.9.3
pycryptodome==3.15.0
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sqlparse==0.4.2
urllib3==1.26.11
